fix: convert BasePrice back to an int

+++ SP23.P03.Web/Features/Trips/Trip.cs
@@ -14,5 +14,5 @@ public class Trip
     public TrainStation? ToStation { get; set; }
     public DateTimeOffset Departure { get; set; }
     public DateTimeOffset Arrival { get; set; }
-    public float BasePrice { get; set; }
+    public int BasePrice { get; set; }
 }
